naming, common core integration tests
