+++ test/minter-suite-minters/MinterSetPrice/MinterSetPriceV0.test.ts
