+++ test/minter-suite-minters/MinterSetPrice/MinterSetPriceV1.PRTNR.test.ts
