+++ test/minter-suite-minters/MinterSetPrice/MinterSetPriceV1.test.ts
