+++ test/minter-suite-minters/MinterSetPriceERC20/MinterSetPriceERC20V0.test.ts
