+++ test/minter-suite-minters/MinterSetPriceERC20/MinterSetPriceERC20V1.PRTNR.test.ts
