+++ test/minter-suite-minters/MinterSetPriceERC20/MinterSetPriceERC20V1.test.ts
