+++ test/minters-pbab/GenArt721Minter_PBAB.test.ts
@@ -1,4 +1,4 @@
-import { GenArt721Minter_PBAB_Base } from "./GenArt721Minter_PBAB_Base.test";
+import { GenArt721Minter_PBAB_Common } from "./GenArt721Minter_PBAB.common";
 
 /**
  * These tests intended to ensure Filtered Minter integrates properly with V1
@@ -7,6 +7,6 @@ import { GenArt721Minter_PBAB_Base } from "./GenArt721Minter_PBAB_Base.test";
 const minter = "GenArt721Minter_PBAB";
 describe(minter, async function () {
   // base tests
-  GenArt721Minter_PBAB_Base(minter);
+  GenArt721Minter_PBAB_Common(minter);
   // no additional tests neeeded for this contract
 });
